Refactoring & implementation of some telemetry/test OpModes to migrate
