+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/DefaultOpMode.java
@@ -4,7 +4,9 @@ package org.firstinspires.ftc.teamcode.common;
  * UserSelection callback representing a default OpMode.
  */
 public final class DefaultOpMode {
-    public DefaultOpMode() {}
+    public DefaultOpMode() {
+    }
+
     @Override
     public String toString() {
         return "DEFAULT";
