+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/DualDeadwheelMecanumDrive.java
@@ -18,7 +18,7 @@ import org.firstinspires.ftc.teamcode.common.roadrunner.util.Encoder;
  *
  * @author Lucas Bubner, 2023
  */
-public class DualDeadwheelMecanumDrive extends Mecanum {
+public class DualDeadwheelMecanumDrive extends MecanumDrive {
     public DualDeadwheelMecanumDrive(@NonNull BunyipsOpMode opMode, DriveConstants constants, MecanumCoefficients mecanumCoefficients, HardwareMap.DeviceMapping<VoltageSensor> voltageSensor, IMU imu, DcMotorEx frontLeft, DcMotorEx backLeft, DcMotorEx frontRight, DcMotorEx backRight, TwoWheelTrackingLocalizerCoefficients localizerCoefficients, Encoder parallel, Encoder perpendicular) {
         super(opMode, constants, mecanumCoefficients, voltageSensor, imu, frontLeft, backLeft, frontRight, backRight);
         setLocalizer(new TwoWheelTrackingLocalizer(localizerCoefficients, parallel, perpendicular, getInstance()));
