+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/MecanumDrive.java
@@ -1,219 +1,213 @@
 package org.firstinspires.ftc.teamcode.common;
 
+import static org.firstinspires.ftc.teamcode.common.Text.round;
+
 import androidx.annotation.NonNull;
 
+import com.acmerobotics.roadrunner.drive.DriveSignal;
+import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.acmerobotics.roadrunner.localization.Localizer;
+import com.acmerobotics.roadrunner.trajectory.Trajectory;
+import com.acmerobotics.roadrunner.trajectory.TrajectoryBuilder;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.util.Range;
+import com.qualcomm.robotcore.hardware.DcMotorEx;
+import com.qualcomm.robotcore.hardware.HardwareMap;
+import com.qualcomm.robotcore.hardware.IMU;
+import com.qualcomm.robotcore.hardware.PIDFCoefficients;
+import com.qualcomm.robotcore.hardware.VoltageSensor;
 
-import java.util.Locale;
+import org.firstinspires.ftc.teamcode.common.roadrunner.drive.DriveConstants;
+import org.firstinspires.ftc.teamcode.common.roadrunner.drive.MecanumCoefficients;
+import org.firstinspires.ftc.teamcode.common.roadrunner.drive.MecanumRoadRunnerDrive;
+import org.firstinspires.ftc.teamcode.common.roadrunner.drive.RoadRunnerDrive;
+import org.firstinspires.ftc.teamcode.common.roadrunner.trajectorysequence.TrajectorySequence;
+
+import java.util.List;
 
 /**
- * Base class for a drive system that uses Mecanum wheels.
- * Includes all the common math done across all Mecanum drive systems.
- * Should be extended instead of implemented directly to allow for additional methods that may be
- * needed as the drive system is developed for a robot.
+ * Wrapper component for the RoadRunner Mecanum Drive, bringing RoadRunner functionality
+ * to the BunyipsOpMode framework.
  *
  * @author Lucas Bubner, 2023
- * @deprecated Deprecated by RoadRunner
  */
-@Deprecated
-public abstract class MecanumDrive extends BunyipsComponent {
+public class MecanumDrive extends BunyipsComponent implements RoadRunnerDrive {
+    private final MecanumRoadRunnerDrive drive;
 
-    private final DcMotor frontLeft;
-    private final DcMotor backLeft;
-    private final DcMotor frontRight;
-    private final DcMotor backRight;
-
-    // Axial translation speeds
-    public double speedX;
-    public double speedY;
-    public double speedR;
-
-    // Store and declare prioritisation when given instruction to calculate motor powers
-    private Priority priority = Priority.NORMALISED;
-
-    protected MecanumDrive(@NonNull BunyipsOpMode opMode, DcMotor frontLeft, DcMotor backLeft, DcMotor frontRight, DcMotor backRight) {
+    public MecanumDrive(@NonNull BunyipsOpMode opMode, DriveConstants constants, MecanumCoefficients mecanumCoefficients, HardwareMap.DeviceMapping<VoltageSensor> voltageSensor, IMU imu, DcMotorEx frontLeft, DcMotorEx backLeft, DcMotorEx frontRight, DcMotorEx backRight) {
         super(opMode);
-        this.frontLeft = frontLeft;
-        this.backLeft = backLeft;
-        this.frontRight = frontRight;
-        this.backRight = backRight;
-        // Critical component, cannot be ignored if null
-        assert frontLeft != null && backLeft != null && frontRight != null && backRight != null;
-    }
-
-    // Setters for the prioritisation of the drive system
-    public void setPriority(Priority priority) {
-        this.priority = priority;
-    }
-
-    public void swapPriority() {
-        priority = priority == Priority.NORMALISED ? Priority.ROTATIONAL : Priority.NORMALISED;
+        drive = new MecanumRoadRunnerDrive(constants, mecanumCoefficients, voltageSensor, imu, frontLeft, frontRight, backLeft, backRight);
+        if (GlobalStorage.contains("lastPose")) {
+            drive.setPoseEstimate((Pose2d) GlobalStorage.get("lastPose"));
+        }
     }
 
     /**
-     * Set a speed at which the Mecanum drive assembly should move using controller input.
-     * The difference is that the vectors will be negated properly for the gamepad.
-     *
-     * @param left_stick_x  X value of the controller
-     * @param left_stick_y  Y value of the controller
-     * @param right_stick_x R value of the controller
-     * @see org.firstinspires.ftc.teamcode.common.Controller#Companion
+     * Cleanup and store the last pose estimate in global storage.
+     * Should be run in BunyipsOpMode.onStop()
      */
-    public void setSpeedUsingController(double left_stick_x, double left_stick_y, double right_stick_x) {
-        speedX = Range.clip(left_stick_x, -1.0, 1.0);
-        speedY = Range.clip(-left_stick_y, -1.0, 1.0);
-        speedR = Range.clip(right_stick_x, -1.0, 1.0);
+    public void teardown() {
+        // Store the last pose estimate in global storage
+        GlobalStorage.put("lastPose", drive.getPoseEstimate());
     }
 
-    /**
-     * Set a speed at which the Mecanum drive assembly should move.
-     *
-     * @param x The speed at which the robot should move in the x direction.
-     *          Positive is right, negative is left.
-     *          Range: -1.0 to 1.0
-     * @param y The speed at which the robot should move in the y direction.
-     *          Positive is forward, negative is backward.
-     *          Range: -1.0 to 1.0
-     * @param r The speed at which the robot will rotate.
-     *          Positive is clockwise, negative is anti-clockwise.
-     *          Range: -1.0 to 1.0
-     */
-    public void setSpeedXYR(double x, double y, double r) {
-        speedX = Range.clip(x, -1.0, 1.0);
-        speedY = Range.clip(y, -1.0, 1.0);
-        speedR = Range.clip(r, -1.0, 1.0);
-    }
-
-    /**
-     * Set a polar speed at which the Mecanum drive assembly should move.
-     *
-     * @param speed             speed at which the motors will operate
-     * @param direction_degrees direction at which the motors will move toward
-     * @param speedR            rotation speed - positive: clockwise
-     */
-    public void setSpeedPolarR(double speed, double direction_degrees, double speedR) {
-        double radians = Math.toRadians(direction_degrees);
-        speedX = Range.clip(speed * Math.cos(radians), -1.0, 1.0);
-        speedY = Range.clip(speed * Math.sin(radians), -1.0, 1.0);
-        this.speedR = Range.clip(speedR, -1.0, 1.0);
-    }
-
-    /**
-     * Update and reflect the speed of the drive system to the actual motors, and
-     * calculate the motor powers based on these variables.
-     */
+    @Override
     public void update() {
-        if (priority == Priority.ROTATIONAL) {
-            rotationalUpdate();
-            getOpMode().addTelemetry(String.format(Locale.getDefault(), "Rotation-priority Mecanum Drive: Forward: %.2f, Strafe: %.2f, Rotate: %.2f", speedX, speedY, speedR));
-            return;
-        }
-
-        // Calculate motor powers
-        double frontLeftPower = speedY + speedX + speedR;
-        double frontRightPower = speedY - speedX - speedR;
-        double backLeftPower = speedY - speedX + speedR;
-        double backRightPower = speedY + speedX - speedR;
-
-        double maxPower = Math.max(Math.abs(frontLeftPower), Math.max(Math.abs(frontRightPower), Math.max(Math.abs(backLeftPower), Math.abs(backRightPower))));
-        // If the maximum number is greater than 1.0, then normalise by that number
-        if (maxPower > 1.0) {
-            frontLeftPower = frontLeftPower / maxPower;
-            frontRightPower = frontRightPower / maxPower;
-            backLeftPower = backLeftPower / maxPower;
-            backRightPower = backRightPower / maxPower;
-        }
-
-        // All powers are reversed when sent to the motors
-        frontLeft.setPower(-frontLeftPower);
-        frontRight.setPower(-frontRightPower);
-        backLeft.setPower(-backLeftPower);
-        backRight.setPower(-backRightPower);
-
-        getOpMode().addTelemetry(String.format(Locale.getDefault(), "Mecanum Drive: X: %.2f, Y: %.2f, R: %.2f", speedX, speedY, speedR));
+        getOpMode().addTelemetry("Localizer: X:%cm Y:%cm %deg",
+                round(Inches.toCM(drive.getPoseEstimate().getX()), 1),
+                round(Inches.toCM(drive.getPoseEstimate().getY()), 1),
+                round(drive.getPoseEstimate().getHeading(), 1));
+        drive.update();
     }
 
-    /**
-     * Immediately stop the drive system.
-     */
-    public void stop() {
-        speedX = 0.0;
-        speedY = 0.0;
-        speedR = 0.0;
-        frontLeft.setPower(0.0);
-        frontRight.setPower(0.0);
-        backLeft.setPower(0.0);
-        backRight.setPower(0.0);
-        update();
+    public MecanumRoadRunnerDrive getInstance() {
+        return drive;
     }
 
-    /**
-     * Set the drive system to brake.
-     */
-    public void setToBrake() {
-        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+    @Override
+    public double getExternalHeading() {
+        return drive.getExternalHeading();
     }
 
-    /**
-     * Set the drive system to float.
-     */
-    public void setToFloat() {
-        backLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.FLOAT);
-        backRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.FLOAT);
-        frontLeft.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.FLOAT);
-        frontRight.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.FLOAT);
+    @Override
+    public void setExternalHeading(double value) {
+        drive.setExternalHeading(value);
     }
 
-    private void rotationalUpdate() {
-        // Calculate translational speeds
-        double[] translationValues = {
-                speedY + speedX,
-                speedY - speedX,
-                speedY - speedX,
-                speedY + speedX
-        };
-
-        double[] rotationValues = {
-                -speedR,
-                speedR,
-                -speedR,
-                speedR
-        };
-
-        double scaleFactor = 1.0;
-        double tmpScale = 1.0;
-
-        // Solve this equation backwards
-        // MotorX = TranslationX * scaleFactor + RotationX
-        // to find scaleFactor that ensures -1 <= MotorX <= 1 and 0 < scaleFactor <= 1
-        for (int i = 0; i < 4; i++) {
-            if (Math.abs(translationValues[i] + rotationValues[i]) > 1) {
-                tmpScale = (1 - rotationValues[i]) / translationValues[i];
-            } else if (translationValues[i] + rotationValues[i] < -1) {
-                tmpScale = (rotationValues[i] - 1) / translationValues[i];
-            }
-            if (tmpScale < scaleFactor) {
-                scaleFactor = tmpScale;
-            }
-        }
-
-        double frontLeftPower = translationValues[0] * scaleFactor + rotationValues[0];
-        double frontRightPower = translationValues[1] * scaleFactor + rotationValues[1];
-        double backLeftPower = translationValues[2] * scaleFactor + rotationValues[2];
-        double backRightPower = translationValues[3] * scaleFactor + rotationValues[3];
-
-        frontLeft.setPower(frontLeftPower);
-        frontRight.setPower(frontRightPower);
-        backLeft.setPower(backLeftPower);
-        backRight.setPower(backRightPower);
+    public void setPowers(double v, double v1, double v2, double v3) {
+        drive.setMotorPowers(v, v1, v2, v3);
     }
 
-    enum Priority {
-        NORMALISED, ROTATIONAL
+    @Override
+    public TrajectoryBuilder trajectoryBuilder(Pose2d startPose) {
+        return drive.trajectoryBuilder(startPose);
+    }
+
+    @Override
+    public TrajectoryBuilder trajectoryBuilder(Pose2d startPose, boolean reversed) {
+        return drive.trajectoryBuilder(startPose, reversed);
+    }
+
+    @Override
+    public TrajectoryBuilder trajectoryBuilder(Pose2d startPose, double startHeading) {
+        return drive.trajectoryBuilder(startPose, startHeading);
+    }
+
+    @Override
+    public void turnAsync(double angle) {
+        drive.turnAsync(angle);
+    }
+
+    @Override
+    public void turn(double angle) {
+        drive.turn(angle);
+    }
+
+    @Override
+    public void followTrajectoryAsync(Trajectory trajectory) {
+        drive.followTrajectoryAsync(trajectory);
+    }
+
+    @Override
+    public void followTrajectory(Trajectory trajectory) {
+        drive.followTrajectory(trajectory);
+    }
+
+    @Override
+    public void followTrajectorySequenceAsync(TrajectorySequence trajectorySequence) {
+        drive.followTrajectorySequenceAsync(trajectorySequence);
+    }
+
+    @Override
+    public void followTrajectorySequence(TrajectorySequence trajectorySequence) {
+        drive.followTrajectorySequence(trajectorySequence);
+    }
+
+    @Override
+    public Pose2d getLastError() {
+        return drive.getLastError();
+    }
+
+    @Override
+    public boolean isBusy() {
+        return drive.isBusy();
+    }
+
+    @Override
+    public Pose2d getPoseEstimate() {
+        return drive.getPoseEstimate();
+    }
+
+    @Override
+    public void setPoseEstimate(Pose2d value) {
+        drive.setPoseEstimate(value);
+    }
+
+    @Override
+    public Pose2d getPoseVelocity() {
+        return drive.getPoseVelocity();
+    }
+
+    @Override
+    public void updatePoseEstimate() {
+        drive.updatePoseEstimate();
+    }
+
+    @Override
+    public void setDriveSignal(DriveSignal driveSignal) {
+        drive.setDriveSignal(driveSignal);
+    }
+
+    @Override
+    public void setDrivePower(Pose2d drivePower) {
+        drive.setDrivePower(drivePower);
+    }
+
+    @Override
+    public void setMode(DcMotor.RunMode runMode) {
+        drive.setMode(runMode);
+    }
+
+    @Override
+    public void setZeroPowerBehavior(DcMotor.ZeroPowerBehavior zeroPowerBehavior) {
+        drive.setZeroPowerBehavior(zeroPowerBehavior);
+    }
+
+    @Override
+    public void setPIDFCoefficients(DcMotor.RunMode runMode, PIDFCoefficients coefficients) {
+        drive.setPIDFCoefficients(runMode, coefficients);
+    }
+
+    @Override
+    public void setWeightedDrivePower(Pose2d drivePower) {
+        drive.setWeightedDrivePower(drivePower);
+    }
+
+    @Override
+    public List<Double> getWheelPositions() {
+        return drive.getWheelPositions();
+    }
+
+    @Override
+    public List<Double> getWheelVelocities() {
+        return drive.getWheelVelocities();
+    }
+
+    @Override
+    public double getRawExternalHeading() {
+        return drive.getRawExternalHeading();
+    }
+
+    @Override
+    public Double getExternalHeadingVelocity() {
+        return drive.getExternalHeadingVelocity();
+    }
+
+    @Override
+    public Localizer getLocalizer() {
+        return drive.getLocalizer();
+    }
+
+    public void setLocalizer(Localizer localizer) {
+        drive.setLocalizer(localizer);
     }
 }
-
