+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/roadrunner/drive/RoadRunnerDrive.java
@@ -24,7 +24,7 @@ import java.util.List;
  *
  * @author Lucas Bubner, 2023
  */
-public interface Drive {
+public interface RoadRunnerDrive {
     static TrajectoryVelocityConstraint getVelocityConstraint(double maxVel, double maxAngularVel, double trackWidth) {
         return new MinVelocityConstraint(Arrays.asList(
                 new AngularVelocityConstraint(maxAngularVel),
