+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/tasks/RoadRunnerTask.java
@@ -5,7 +5,7 @@ import androidx.annotation.NonNull;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 
 import org.firstinspires.ftc.teamcode.common.BunyipsOpMode;
-import org.firstinspires.ftc.teamcode.common.Mecanum;
+import org.firstinspires.ftc.teamcode.common.MecanumDrive;
 import org.firstinspires.ftc.teamcode.common.roadrunner.trajectorysequence.TrajectorySequence;
 
 /**
@@ -14,18 +14,18 @@ import org.firstinspires.ftc.teamcode.common.roadrunner.trajectorysequence.Traje
  * @author Lucas Bubner, 2023
  */
 public class RoadRunnerTask extends Task {
-    private final Mecanum drive;
+    private final MecanumDrive drive;
 
     private Trajectory trajectory;
     private TrajectorySequence trajectorySequence;
 
-    public RoadRunnerTask(@NonNull BunyipsOpMode opMode, double time, Mecanum drive, Trajectory trajectory) {
+    public RoadRunnerTask(@NonNull BunyipsOpMode opMode, double time, MecanumDrive drive, Trajectory trajectory) {
         super(opMode, time);
         this.drive = drive;
         this.trajectory = trajectory;
     }
 
-    public RoadRunnerTask(@NonNull BunyipsOpMode opMode, double time, Mecanum drive, TrajectorySequence trajectorySequence) {
+    public RoadRunnerTask(@NonNull BunyipsOpMode opMode, double time, MecanumDrive drive, TrajectorySequence trajectorySequence) {
         super(opMode, time);
         this.drive = drive;
         this.trajectorySequence = trajectorySequence;
