+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/glados/components/GLaDOSPOVDriveCore.java
@@ -5,9 +5,9 @@ import androidx.annotation.NonNull;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
 import org.firstinspires.ftc.teamcode.common.BunyipsOpMode;
-import org.firstinspires.ftc.teamcode.common.MecanumDrive;
+import org.firstinspires.ftc.teamcode.common.StandardMecanumDrive;
 
-public class GLaDOSPOVDriveCore extends MecanumDrive {
+public class GLaDOSPOVDriveCore extends StandardMecanumDrive {
     public GLaDOSPOVDriveCore(@NonNull BunyipsOpMode opMode, DcMotor frontLeft, DcMotor backLeft, DcMotor frontRight, DcMotor backRight) {
         super(opMode, frontLeft, backLeft, frontRight, backRight);
     }
