+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/jerry/components/JerryDrive.kt
@@ -2,7 +2,7 @@ package org.firstinspires.ftc.teamcode.jerry.components
 
 import com.qualcomm.robotcore.hardware.DcMotorEx
 import org.firstinspires.ftc.teamcode.common.BunyipsOpMode
-import org.firstinspires.ftc.teamcode.common.MecanumDrive
+import org.firstinspires.ftc.teamcode.common.StandardMecanumDrive
 
 /**
  * Jerry robot drivetrain operation module.
@@ -11,4 +11,4 @@ class JerryDrive(
     opMode: BunyipsOpMode,
     bl: DcMotorEx, br: DcMotorEx,
     fl: DcMotorEx, fr: DcMotorEx
-) : MecanumDrive(opMode, fl, bl, fr, br)
+) : StandardMecanumDrive(opMode, fl, bl, fr, br)
