+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/jerry/components/JerryPolarDrive.kt
@@ -2,7 +2,7 @@ package org.firstinspires.ftc.teamcode.jerry.components
 
 import com.qualcomm.robotcore.hardware.DcMotorEx
 import org.firstinspires.ftc.teamcode.common.BunyipsOpMode
-import org.firstinspires.ftc.teamcode.common.FieldCentricMecanumDrive
+import org.firstinspires.ftc.teamcode.common.FieldCentricStandardMecanumDrive
 import org.firstinspires.ftc.teamcode.common.IMUOp
 import org.firstinspires.ftc.teamcode.common.RelativePose2d
 
@@ -11,4 +11,4 @@ class JerryPolarDrive(
     bl: DcMotorEx, br: DcMotorEx,
     fl: DcMotorEx, fr: DcMotorEx,
     imu: IMUOp, startingDirection: RelativePose2d
-) : FieldCentricMecanumDrive(opMode, fl, bl, fr, br, imu, true, startingDirection)
+) : FieldCentricStandardMecanumDrive(opMode, fl, bl, fr, br, imu, true, startingDirection)
