+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wheatley/teleop/WheatleyTeleOp.java
@@ -49,8 +49,6 @@ public class WheatleyTeleOp extends BunyipsOpMode {
             suspender = new WheatleyManagementRail(this, config.susMotor, config.susServo);
         if (NullSafety.assertComponentArgs(this, Cannon.class, config.pl))
             cannon = new Cannon(this, config.pl);
-
-        // TODO: Review systems for operator ease of use
     }
 
     @Override
